move vtk material in submodule
